requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
dnspython==2.7.0
idna==3.10
pycparser==2.22
python-nmap==0.7.1
requests==2.32.3
soupsieve==2.7
typing_extensions==4.13.2
urllib3==2.4.0
validators==0.35.0